# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: folicular
  description: |
    Backend service for Luteal: anonymous identity, offline-first delta sync,
    research-backed canonical data model, clearly labeled cycle estimates.

    This document is the contract's single source of truth. The Go server is
    validated against it (internal/contract tests); the Android client
    generates its DTOs from it. Prose documentation lives in docs/api.md.

    Conventions: dates are ISO-8601 calendar dates (format: date); instants
    are RFC 3339 UTC (format: date-time). Synchronized record IDs are
    client-generated UUIDv7 strings. Errors are RFC 9457 problem+json.
  version: 1.0.0
  contact:
    name: Luteal
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/devices/{deviceID}:
    delete:
      tags: [auth]
      operationId: revokeDevice
      summary: Revoke a device token
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      responses:
        "204": { description: Device revoked }
        "404": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 256-bit device token issued at registration (ltok_...)
  parameters:
    DeviceID:
      name: deviceID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    UUID:
      type: string
      format: uuid
      example: 019832e0-6c14-7000-8000-000000000001
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...